fix(router): extract pooled feed route to keep app_router under size ceiling
The redirect fix grew app_router.dart past its frozen 800+-line ceiling (#4339
ratchet). Move the fullscreen feed route's builder + redirect helper into
pooled_fullscreen_feed_route.dart; app_router.dart now shrinks (1562 -> 1534).
